integration: move test_refresh_metadata_after_schema_agreement
This obviously belongs to `metadata/configuration.rs` category.
