+++ scylla/tests/integration/metadata/configuration.rs
@@ -1,9 +1,11 @@
 //! Tests that the driver enforces the provided custom metadata fetching timeout
 //! iff ScyllaDB is the target node (else ignores the custom timeout).
 
+use crate::utils::{create_new_session_builder, unique_keyspace_name, PerformDDL as _};
 use crate::utils::{setup_tracing, test_with_3_node_cluster};
 use scylla::client::session::Session;
 use scylla::client::session_builder::SessionBuilder;
+use scylla::cluster::metadata::{ColumnType, NativeType, UserDefinedType};
 use scylla_proxy::{
     Condition, Reaction as _, RequestFrame, RequestOpcode, RequestReaction, RequestRule,
     ShardAwareness,
@@ -210,4 +212,36 @@ async fn test_custom_metadata_timeouts() {
         Err(ProxyError::Worker(WorkerError::DriverDisconnected(_))) => (),
         Err(err) => panic!("{}", err),
     }
+}
+
+#[tokio::test]
+async fn test_refresh_metadata_after_schema_agreement() {
+    let session = create_new_session_builder().build().await.unwrap();
+
+    let ks = unique_keyspace_name();
+    session.ddl(format!("CREATE KEYSPACE IF NOT EXISTS {} WITH REPLICATION = {{'class' : 'NetworkTopologyStrategy', 'replication_factor' : 1}}", ks)).await.unwrap();
+    session.use_keyspace(ks.clone(), false).await.unwrap();
+
+    session
+        .ddl("CREATE TYPE udt (field1 int, field2 uuid, field3 text)")
+        .await
+        .unwrap();
+
+    let cluster_state = session.get_cluster_state();
+    let keyspace_metadata = cluster_state.get_keyspace(ks.as_str());
+    assert_ne!(keyspace_metadata, None);
+
+    let udt = keyspace_metadata.unwrap().user_defined_types.get("udt");
+    assert_eq!(
+        udt,
+        Some(&Arc::new(UserDefinedType {
+            keyspace: ks.into(),
+            name: "udt".into(),
+            field_types: Vec::from([
+                ("field1".into(), ColumnType::Native(NativeType::Int)),
+                ("field2".into(), ColumnType::Native(NativeType::Uuid)),
+                ("field3".into(), ColumnType::Native(NativeType::Text))
+            ])
+        }))
+    );
 }
